feat(remediate): tool-owned command shape for transitive pins, plus shared seam groundwork
Two ecosystems arriving in wave 2 (go, cargo) own their manifest and
lockfile formats: hand-editing go.mod or Cargo.lock is the corruption
class the seam exists to refuse, and the honest pin is the tool's own
command (go get pkg@version, cargo update -p pkg --precise version).
The pin plan therefore gains a third, additive member: a command plan
carrying the argv, the files it rewrites, and the revert prose. The
executor runs it at the owning root through the injected bounded exec
and skips the separate lock resync (the one tool command leaves the
tree consistent); edit plans keep the exact behavior they had.

Groundwork moves, no behavior change:
- the concrete-semver shape and the default pin grammar move from the
  recipe tier onto the seam itself (beside PinVersionScheme), so a pack
  composing on the semver order (go's v-prefixed grammar) reads the one
  comparator; the recipe tier re-exports them.
- the line-level TOML table/key scans move from python-remediation.ts
  into languages/toml-text.ts (the rust pack needs the identical scans
  for its direct-dependency refusal); the table-name escape hardens to
  full regex escaping since names now arrive from tomlTableNames too.
